Easy sprite demo fix + doc tweaks

+++ src/cute_string.cpp
@@ -336,7 +336,7 @@ uint64_t cf_stohex(const char* s)
 {
 	if (!CF_STRNCMP(s, "#", 1)) s += 1;
 	if (!CF_STRNCMP(s, "0x", 2)) s += 2;
-	int len = CF_STRLEN(s);
+	int len = (int)CF_STRLEN(s);
 	if (len != 6 && len != 8) return 0;
 	char* end;
 	uint64_t result = CF_STRTOLL(s, &end, 16);
@@ -559,22 +559,22 @@ char* cf_string_append_UTF8_impl(char *s, int codepoint)
 
 const char* cf_decode_UTF8(const char* s, int* codepoint)
 {
-    unsigned char c = *s++;
-    int extra = 0;
-    int min = 0;
-    *codepoint = 0;
-         if (c >= 0xF0) { *codepoint = c & 0x07; extra = 3; min = 0x10000; }
-    else if (c >= 0xE0) { *codepoint = c & 0x0F; extra = 2; min = 0x800; }
-    else if (c >= 0xC0) { *codepoint = c & 0x1F; extra = 1; min = 0x80; }
-    else if (c >= 0x80) { *codepoint = 0xFFFD; }
-    else *codepoint = c;
-    while (extra--) {
-        c = *s++;
-        if ((c & 0xC0) != 0x80) { *codepoint = 0xFFFD; }
-        if (*codepoint != 0xFFFD) { *codepoint = ((*codepoint) << 6) | (c & 0x3F); }
-    }
-    if (*codepoint < min) *codepoint = 0xFFFD;
-    return s;
+	unsigned char c = *s++;
+	int extra = 0;
+	int min = 0;
+	*codepoint = 0;
+			if (c >= 0xF0) { *codepoint = c & 0x07; extra = 3; min = 0x10000; }
+	else if (c >= 0xE0) { *codepoint = c & 0x0F; extra = 2; min = 0x800; }
+	else if (c >= 0xC0) { *codepoint = c & 0x1F; extra = 1; min = 0x80; }
+	else if (c >= 0x80) { *codepoint = 0xFFFD; }
+	else *codepoint = c;
+	while (extra--) {
+		c = *s++;
+		if ((c & 0xC0) != 0x80) { *codepoint = 0xFFFD; }
+		if (*codepoint != 0xFFFD) { *codepoint = ((*codepoint) << 6) | (c & 0x3F); }
+	}
+	if (*codepoint < min) *codepoint = 0xFFFD;
+	return s;
 }
 
 const uint16_t* cf_decode_UTF16(const uint16_t* s, int* codepoint)
